@@ -1,40 +1,37 @@
 {
   "name": "my-component-library-pobolov",
-  "version": "1.0.0",
+  "version": "1.0.1-dev",
+  "private": true,
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "files": ["dist"],
   "scripts": {
+    "start": "webpack serve --config webpack.config.js --open",
     "build": "webpack --config webpack.config.js && tsc",
     "test": "jest",
     "lint": "eslint ."
   },
-  "peerDependencies": {
-    "react": "^18 || ^19",
-    "react-dom": "^18 || ^19"
+  "dependencies": {
+    "@eslint/js": "^9.37.0",
+    "eslint-plugin-react": "^7.37.5",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "typescript-eslint": "^8.46.0"
   },
   "devDependencies": {
     "@babel/core": "^7.28.4",
     "@babel/preset-env": "^7.28.3",
     "@babel/preset-react": "^7.27.1",
     "@babel/preset-typescript": "^7.27.1",
-    "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.0",
-    "@testing-library/user-event": "^14.6.1",
-    "@types/jest": "^30.0.0",
-    "@types/react": "^19.1.15",
-    "@types/react-dom": "^19.1.9",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.1",
     "babel-loader": "^10.0.0",
     "clean-webpack-plugin": "^4.0.0",
     "css-loader": "^7.1.2",
-    "jest": "^30.2.0",
-    "jest-environment-jsdom": "^30.2.0",
+    "html-webpack-plugin": "^5.6.4",
     "style-loader": "^4.0.0",
-    "ts-jest": "^29.4.4",
-    "ts-loader": "^9.5.4",
     "typescript": "^5.9.2",
     "webpack": "^5.101.3",
-    "webpack-cli": "^6.0.1"
-  },
-  "dependencies": {}
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.2"
+  }
 }
