@@ -1,3 +1,7 @@
-module.exports = new Proxy({}, {
+// module.exports = new Proxy({}, {
+//     get: (target, key) => key,
+// });
+
+export default new Proxy({}, {
     get: (target, key) => key,
 });
